==> 2_CSV headlines collector and filtering.ipynb <==
# Collecting and filtering received headlines

## Collect Headline files

import pandas as pd
import numpy as np
import glob
import datetime
from datetime import time


#path = r'C:\DRO\DCL_rawdata_files' # use your path
path = '' # use your path
all_files = glob.glob("4_WRDS_SP 500 Headlines completed/Single Files/*.csv")

li = []

# Load all files
for filename in all_files:
    df = pd.read_csv(filename)
    li.append(df)

frame = pd.concat(li, axis=0, ignore_index=False)

frame

# convert versionCreated to datetime
frame.versionCreated =pd.to_datetime(frame.versionCreated)
frame

# Set index to version created without timestamp
frame.set_index(frame['versionCreated'].dt.normalize(),inplace=True)

## Filter Headlines

# Drop original 'index' timestamp
frame.drop(columns=['Unnamed: 0'],inplace=True)
frame

sum(frame["RIC"]=="APA.O")

# Change APA.0 to APA (fix that a different Ticker was used for same company)
frame["RIC"].replace({"APA.O": "APA"}, inplace=True)

sum(frame["RIC"]=="APA.O")

# Sort dataframe by index
frame = frame.sort_index()

# Remove duplicates
frame= frame.drop_duplicates()
frame

frame[frame.sourceCode== "NS:EDG"]

frame[frame.sourceCode== "NS:GLFILE"]

frame[frame.sourceCode== "NS:TRANS"]

# Filter out specific news sources
frame=frame[frame.sourceCode!= "NS:EDG"]
frame=frame[frame.sourceCode!= "NS:TRANS"]
frame=frame[frame.sourceCode!= "NS:GLFILE"]
frame

# remove headlines with wrong tagging of CBOE.K unitl end of June 2019
frame[((frame.sourceCode== "NS:PUBT") & (frame.RIC=="CBOE.K") & (frame.index<"2020-06-30"))]

# remove headlines with wrong tagging of CBOE.K unitl end of June 2019
frame=frame[~((frame.sourceCode== "NS:PUBT") & (frame.RIC=="CBOE.K") & (frame.index<"2020-06-30"))]
frame

frame[frame.RIC== "MCO"]

# remove headlines for Moodys
frame=frame[frame.RIC!= "MCO"]
frame

frame.index.max()

# Filter out February / July news
frame=frame.loc['2019-03-01':'2020-06-30']
frame

frame.tail(10)

# Store frame into dataframe
frame.to_csv("4_WRDS_SP 500 Headlines completed/2019_03_01_to_2020_06_30_Headlines_SP500.csv")

# create flag column that concatenates text, sourcecode and RIC
frame["Duplicate_Flag"]= (frame.text+frame.sourceCode+frame.RIC).str.lower()
frame

# find day with duplicates headlines
frame.loc["2020-06-01"].sort_values("text",ascending=False)

# rename index label
frame.index.rename('versionCreated.1',inplace=True)
frame

# sort by version created
frame = frame.sort_values(by=frame.columns[0])
frame

frame.drop_duplicates("Duplicate_Flag",inplace=True)
frame

# check again day with duplicates headlines
frame.loc["2020-06-01"].sort_values("text",ascending=False)

frame=frame.drop("Duplicate_Flag",axis=1)
frame

# Check how many news are on weekends
import holidays

us_holidays = holidays.UnitedStates()

from datetime import date
# As the expand parameter of uk_holidays= True, the list of holidays will expand
# for each year that is checked. Therfore, we generate a loop to effectivley initialise all 
# holidays from 2000- 2020 in the holiday object
for x in range (2019,2020):
  date(x, 1, 1) in us_holidays

# Check length of holiday object
len(us_holidays)

frame["Holiday_Flag"]=frame.index.isin(us_holidays)
frame

frame[frame.Holiday_Flag==True]

# Filter out holidays
frame=frame[frame.Holiday_Flag==False]
frame

dict_days={0:"Mon",
          1:"Tue",
          2:"Wed",
          3:"Thu",
          4:"Fri",
          5:"Sat",
          6:"Sun"}


# Insert controll variable for the day of the week (0=Monday, 6=Sunday)
frame["Weekday"]=pd.to_datetime(frame.index).dayofweek.map(dict_days)
frame

frame[frame.Weekday.isin(["Sat","Sun"])]

frame=frame[~frame.Weekday.isin(["Sat","Sun"])]
frame

frame=frame.drop(columns=["Holiday_Flag","Weekday"])
frame

print(662739-8772)

# Store frame into dataframe
frame.to_csv("4_WRDS_SP 500 Headlines completed/2019_03_01_to_2020_06_30_Headlines_SP500_filt.csv")

### Analysis of news headlines

frame["storyId"].value_counts() 

# unique news
frame["storyId"].nunique()

# Create dataframe containing unique storyIds
news_Ids=frame[["storyId","sourceCode"]]
print(news_Ids.shape)
news_Ids=news_Ids.drop_duplicates()
print(news_Ids.shape)

# Store dataframe containing only unique headlines
news_Ids.to_csv("4_WRDS_SP 500 Headlines completed/2019_03_01_to_2020_04_30_unique_HeadlineIds_SP100.csv")

# Rough monthly estimate
frame["storyId"].nunique()/13

# Print latest news date
max(frame.versionCreated)

# average news frequency per company per day
frame["storyId"].nunique()/505/370

frame["month_year"]=frame.versionCreated.dt.strftime('%Y-%m')
frame["month_year_day"]=frame.versionCreated.dt.strftime('%Y-%m-%d')



# Generate pivot table of sum of news per day per company
piv_table_news = pd.pivot_table(frame,index=['RIC'],columns=[frame.month_year],\
               values=["storyId"],\
               aggfunc=['count'],fill_value=0)

# Display the table
piv_table_news

# Store in excel
piv_table_news.to_excel("4_WRDS_SP 500 Headlines completed/Overview News per Company and month.xlsx")

Analysis Daily averages

frame

# Analysis daily averages

# Generate pivot table of sum of news per day per company
piv_table_news_daily = pd.pivot_table(frame,index=['RIC'],columns=[frame.month_year_day],\
               values=["storyId"],\
               aggfunc=['count'],fill_value=0)

# Display the table
piv_table_news_daily

piv_table_news_daily[piv_table_news_daily.index=="CBOE.K"].to_excel("check_CBOE.K_new_vol.xlsx")
piv_table_news[piv_table_news.index=="CBOE.K"].to_excel("check_CBOE.K_new_vol_monthly.xlsx")

frame[frame.RIC=="CBOE.K"].reset_index()[["month_year_day","text","sourceCode","RIC"]].to_excel("CBOE.K News.xlsx")

#check only news for PUBT 
frame[frame.sourceCode=="NS:PUBT"].reset_index()[["month_year_day","text","sourceCode","RIC"]].to_excel("PUBT News.xlsx")

piv_table_news_daily_sum= pd.DataFrame()
piv_table_news_daily_sum["Mean"]=piv_table_news_daily.mean(axis=1)
piv_table_news_daily_sum["Median"]=piv_table_news_daily.median(axis=1)
piv_table_news_daily_sum["Max"]=piv_table_news_daily.max(axis=1)
piv_table_news_daily_sum["Count_days"]=piv_table_news_daily.count(axis=1)
piv_table_news_daily_sum

# calculate days with no articles
piv_table_news_daily_sum["days_with_no_articles"]=(piv_table_news_daily == 0).sum(axis=1)
piv_table_news_daily_sum

# calculate % of days with no articles
piv_table_news_daily_sum["days_with_no_articles_PERC"]=piv_table_news_daily_sum["days_with_no_articles"]/piv_table_news_daily_sum["Count_days"]
piv_table_news_daily_sum

# Store in excel
piv_table_news_daily_sum.to_excel("4_WRDS_SP 500 Headlines completed/Overview News count per Company and day.xlsx")

Analysis by Source

frame["month_year"]=frame.versionCreated.dt.strftime('%Y-%m')

# Generate pivot table of sum news per author
piv_table_authors = pd.pivot_table(frame,index=['sourceCode'],columns=[frame.month_year],\
               values=["storyId"],\
               aggfunc=['count'],fill_value=0)

# Display the table
piv_table_authors

# Store in excel
piv_table_authors.to_excel("4_WRDS_SP 500 Headlines completed/Overview News per Author and month.xlsx")

# Check how many news are on weekends
import holidays

us_holidays = holidays.UnitedStates()

from datetime import date
# As the expand parameter of uk_holidays= True, the list of holidays will expand
# for each year that is checked. Therfore, we generate a loop to effectivley initialise all 
# holidays from 2000- 2020 in the holiday object
for x in range (2019,2020):
  date(x, 1, 1) in us_holidays

# Check length of holiday object
len(us_holidays)

frame["Holiday_Flag"]=frame.index.isin(us_holidays)
frame

frame.loc["2019-05-27"]

frame[frame.Holiday_Flag==True]

for ptr in holidays.UnitedStates(years = 2019).items(): 
    print(ptr) 

5,177 news on holidays

dict_days={0:"Mon",
          1:"Tue",
          2:"Wed",
          3:"Thu",
          4:"Fri",
          5:"Sat",
          6:"Sun"}


# Insert controll variable for the day of the week (0=Monday, 6=Sunday)
frame["Weekday"]=pd.to_datetime(frame.index).dayofweek.map(dict_days)
frame

frame[frame.Weekday.isin(["Sat","Sun"])]

34,123 news are on weekends.

**Summary:**

Holliday and Weekend news account for

34,123 + 5,177 = 39,300 news

of 702,039 news articles

39300/702039= 0,5597 = 5.6% of all news
